+++ strands-py/src/strands/multiagent/graph.py
@@ -496,7 +496,7 @@ class Graph(MultiAgentBase):
         run_async(lambda: self.hooks.invoke_callbacks_async(MultiAgentInitializedEvent(self)))
 
     def add_hook(
-        self, callback: HookCallback, event_type: type | list[type] | None = None, *, order: int = HookOrder.DEFAULT
+        self, callback: HookCallback, event_type: type | list[type] | None = None, *, order: float = HookOrder.DEFAULT
     ) -> None:
         """Register a hook callback with the graph.
